# openapi-spec.yaml
openapi: 3.0.0
info:
  title: Clare and Me API
  description: API for Clare and Me mental health support chatbot
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /retrieveContext/{userId}:
    get:
      summary: Retrieve context for a user
      tags:
        - Context Management
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  context:
                    type: object
        '401':
          description: Unauthorized
  /updateContext:
    post:
      summary: Update context for a user
      tags:
        - Context Management
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContextDto'
      responses:
        '201':
          description: Context updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateContextDto:
      type: object
      properties:
        userId:
          type: string
        context:
          type: object
      required:
        - userId
        - context